+++ src/number/index.js
@@ -1,7 +1,9 @@
 'use strict'
 
 import pad from './pad'
+import toNumber from './toNumber'
 
 export {
     pad,
+    toNumber,
 }