+++ src/string/index.js
@@ -6,6 +6,15 @@ import startsWith from './startsWith'
 
 export const reTrim = /^\s+|\s+$/g
 
+/** Used to detect binary string values. */
+export const reIsBinary = /^0b[01]+$/i
+
+/** Used to detect octal string values. */
+export const reIsOctal = /^0o[0-7]+$/i
+
+/** Used to detect bad signed hexadecimal string values. */
+export const reIsBadHex = /^[-+]0x[0-9a-f]+$/i
+
 /**
  * This function trim string
  *